requirements.txt
gradio
requests
beautifulsoup4
pandas
chromadb
langchain-groq
python-dotenv
